+++ database/models/Guild.py
@@ -72,6 +72,7 @@ class Guild(Model):
         'default_channel_state': 'bool',
         'default_member_state': 'bool',
         'default_role_state': 'bool',
+        'silent' : 'bool'  #Add in Database,
     }
 
     @has_many('id', 'guild_id')
