+++ requirements.txt
@@ -1,4 +1,3 @@
-git+https://github.com/Kyrela/discore
 python-i18n~=0.3.9
 psutil~=6.1.1
 requests~=2.32.0
@@ -7,3 +6,4 @@ pymysql~=1.1.1
 git+https://github.com/Kyrela/discord-markdown-ast-parser
 git+https://github.com/null8626/python-sdk@patch-1
 aiohttp~=3.11.14
+git+https://github.com/Kyrela/discore@1659f179536814b5f672db36821d044c52fa09b5
